Plan 9 from Bell Labs Fourth Edition 2007-01-22

+++ sys/src/9/port/devenv.c
@@ -262,7 +262,7 @@ static long
 envwrite(Chan *c, void *a, long n, vlong off)
 {
 	char *s;
-	int vend;
+	ulong vend;
 	Egrp *eg;
 	Evalue *e;
 	ulong offset = off;
